+++ examples/trials/pix2pix-pytorch/.gitignore
@@ -1,2 +1,5 @@
 # datasets
 data/
+
+# pix2pix library
+pix2pixlib/
