--- examples/trials/pix2pix-pytorch/config.yml
@@ -1,24 +1,0 @@
-authorName: default
-experimentName: example_pix2pix
-trialConcurrency: 1
-maxExecDuration: 10h
-maxTrialNum: 1
-#choice: local, remote, pai
-trainingServicePlatform: local
-searchSpacePath: search_space.json
-#choice: true, false
-useAnnotation: false
-tuner:
-  #choice: TPE, Random, Anneal, Evolution, BatchTuner, MetisTuner, GPTuner
-  #SMAC (SMAC should be installed through nnictl)
-  builtinTunerName: TPE
-  classArgs:
-    #choice: maximize, minimize
-    optimize_mode: minimize
-trial:
-  command: python3 pix2pix.py
-  codeDir: .
-  gpuNum: 1
-localConfig:
-  useActiveGpu: true
-  gpuIndices: '0'
